% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokingMouse-package.R
\docType{package}
\name{smokingMouse-package}
\alias{smokingMouse-package}
\alias{smokingMouse}
\title{Provides access to raw and processed data of the smokingMouse project}
\description{
This is an ExperimentHub package that provides access to the raw and processed data generated at the gene, exon, tx and jx levels,
in all phases of the smoking mouse project. See https://github.com/LieberInstitute/smokingMouse_Indirects.
}
\author{
Daianna Gonzalez-Padilla
}
