% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokingMouse-package.R
\docType{package}
\name{smokingMouse-package}
\alias{smokingMouse-package}
\alias{smokingMouse}
\title{Provides access to data of the smokingMouse project}
\description{
This is an ExperimentHub package that provides access to the data at the gene, 
exon, transcript and junction level, used in the analyses of the smokingMouse 
project. See https://github.com/LieberInstitute/smokingMouse_Indirects.
License: Artistic-2.0
}
\author{
Daianna Gonzalez-Padilla
}
