% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokingMouse-package.R
\docType{package}
\name{smokingMouse-package}
\alias{smokingMouse-package}
\alias{smoking-nicotine-mouse}
\title{Get access to the smoking-nicotine-mouse LIBD project data}
\description{
This is an ExperimentHub package that provides access to the data generated and analyzed in the \href{https://github.com/LieberInstitute/smoking-nicotine-mouse/}{smoking-nicotine-mouse} LIBD project.
License: Artistic-2.0
}
\author{
Daianna Gonzalez-Padilla
}
